requirements/testing.txt
coverage
pytest-localserver
pytest-cov==3.0.0
selenium<4.0
